+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.3.0](https://github.com/gumlet/python-sdk/compare/v0.2.0...v0.3.0) (2026-09-21)
+
+
+### Features
+
+* **api:** add operation liveStreamAnalytics.usage ([ac60a30](https://github.com/gumlet/python-sdk/commit/ac60a305166b5bcd069f543f67c4a40308d71331))
+
+
+### Chores
+
+* **api:** regenerate SDK ([02bd3cb](https://github.com/gumlet/python-sdk/commit/02bd3cb58ad9d5bd2f46b0baf29ef9e29e5d4b7a))
+* **api:** update generated SDK content ([04b0f2e](https://github.com/gumlet/python-sdk/commit/04b0f2eb66043f416c7779a5507efa648258287a))
+* **api:** update generated SDK content ([82f24bf](https://github.com/gumlet/python-sdk/commit/82f24bf8486f195dfc974657c8bb7d40419b1a6d))
+
 ## [0.2.0](https://github.com/gumlet/python-sdk/compare/v0.1.0...v0.2.0) (2026-09-18)